+++ arcade/cinematic/camera_2D.py
@@ -1,0 +1,6 @@
+# TODO
+class Camera2D:
+    """
+    A simple orthographic camera. Similar to SimpleCamera, but takes better advantage of the new data structures.
+    As the Simple Camera is depreciated any new project should use this camera instead.
+    """
